refactor: refactor Console & MiniConsole
