+++ src/components/panels/MiniconsolePanel.vue
@@ -10,7 +10,7 @@
             <v-btn icon tile @click="clearConsole">
                 <v-icon small>{{ mdiTrashCan }}</v-icon>
             </v-btn>
-            <command-help-modal :in-toolbar="true" @onCommand="gcode = $event" />
+            <command-help-modal :in-toolbar="true" @onCommand="commandClick($event)" />
             <v-menu
                 :offset-y="true"
                 :close-on-content-click="false"
@@ -62,28 +62,7 @@
         </template>
         <div class="d-flex flex-column">
             <v-card-text :class="consoleDirection === 'table' ? 'order-1' : 'order-2'">
-                <v-textarea
-                    ref="gcodeCommandField"
-                    v-model="gcode"
-                    :items="items"
-                    :label="$t('Panels.MiniconsolePanel.SendCode')"
-                    solo
-                    class="gcode-command-field"
-                    autocomplete="off"
-                    no-resize
-                    auto-grow
-                    :rows="rows"
-                    hide-details
-                    outlined
-                    dense
-                    :prepend-icon="isTouchDevice ? mdiChevronDoubleRight : ''"
-                    :append-icon="mdiSend"
-                    @keydown.enter.prevent.stop="doSend"
-                    @keyup.up="onKeyUp"
-                    @keyup.down="onKeyDown"
-                    @keydown.tab="getAutocomplete"
-                    @click:prepend="getAutocomplete"
-                    @click:append="doSend"></v-textarea>
+                <console-textarea ref="gcodeCommandField" />
             </v-card-text>
             <v-card-text :class="(consoleDirection === 'table' ? 'order-2' : 'order-1') + ' pa-0'">
                 <v-row>
@@ -107,14 +86,14 @@
 </template>
 
 <script lang="ts">
-import { strLongestEqual, reverseString } from '@/plugins/helpers'
-import { Component, Mixins, Watch } from 'vue-property-decorator'
+import { Component, Mixins, Ref, Watch } from 'vue-property-decorator'
 import BaseMixin from '@/components/mixins/base'
-import { CommandHelp, VTextareaType } from '@/store/printer/types'
 import ConsoleTable from '@/components/console/ConsoleTable.vue'
 import Panel from '@/components/ui/Panel.vue'
-import { mdiChevronDoubleRight, mdiCog, mdiConsoleLine, mdiSend, mdiTrashCan } from '@mdi/js'
+import { mdiCog, mdiConsoleLine, mdiTrashCan } from '@mdi/js'
 import CommandHelpModal from '@/components/CommandHelpModal.vue'
+import ConsoleMixin from '@/components/mixins/console'
+import ConsoleTextarea from '@/components/inputs/ConsoleTextarea.vue'
 
 @Component({
     components: {
@@ -123,31 +102,13 @@ import CommandHelpModal from '@/components/CommandHelpModal.vue'
         CommandHelpModal,
     },
 })
-export default class MiniconsolePanel extends Mixins(BaseMixin) {
+export default class MiniconsolePanel extends Mixins(BaseMixin, ConsoleMixin) {
     mdiTrashCan = mdiTrashCan
     mdiConsoleLine = mdiConsoleLine
     mdiCog = mdiCog
-    mdiSend = mdiSend
-    mdiChevronDoubleRight = mdiChevronDoubleRight
 
-    declare $refs: {
-        gcodeCommandField: VTextareaType
-        console: ConsoleTable
-        miniConsoleScroll: any
-    }
-
-    private gcode = ''
-    private lastCommandNumber: number | null = null
-    private items = []
-    private cmdListSearch: string | null = null
-
-    get helplist(): CommandHelp[] {
-        return this.$store.state.printer.helplist ?? []
-    }
-
-    get consoleDirection() {
-        return this.$store.state.gui.console.direction ?? 'table'
-    }
+    @Ref() readonly miniConsoleScroll!: any
+    @Ref() readonly gcodeCommandField!: typeof ConsoleTextarea
 
     get consoleHeight() {
         return this.$store.state.gui.console.height ?? 300
@@ -171,160 +132,8 @@ export default class MiniconsolePanel extends Mixins(BaseMixin) {
         if (newVal) this.scrollToBottom()
     }
 
-    clearConsole() {
-        this.$store.dispatch('gui/console/clear')
-    }
-
-    get hideWaitTemperatures(): boolean {
-        return this.$store.state.gui.console.hideWaitTemperatures
-    }
-
-    set hideWaitTemperatures(newVal) {
-        this.$store.dispatch('gui/saveSetting', { name: 'console.hideWaitTemperatures', value: newVal })
-    }
-
-    get hideTlCommands(): boolean {
-        return this.$store.state.gui.console.hideTlCommands
-    }
-
-    set hideTlCommands(newVal) {
-        this.$store.dispatch('gui/saveSetting', { name: 'console.hideTlCommands', value: newVal })
-    }
-
-    get customFilters(): any[] {
-        return this.$store.state.gui.console.consolefilters
-    }
-
-    get rows(): number {
-        return this.gcode?.split('\n').length ?? 1
-    }
-
-    get lastCommands(): string[] {
-        return this.$store.state.gui.gcodehistory.entries ?? []
-    }
-
-    get autoscroll(): boolean {
-        return this.$store.state.gui.console.autoscroll ?? true
-    }
-
-    set autoscroll(newVal) {
-        this.$store.dispatch('gui/saveSetting', { name: 'console.autoscroll', value: newVal })
-    }
-
-    get rawOutput(): boolean {
-        return this.$store.state.gui.console.rawOutput ?? false
-    }
-
-    set rawOutput(newVal) {
-        this.$store.dispatch('gui/saveSetting', { name: 'console.rawOutput', value: newVal })
-    }
-
     commandClick(msg: string): void {
-        this.gcode = msg
-
-        this.$nextTick(() => {
-            this.$refs.gcodeCommandField.focus()
-        })
-    }
-
-    doSend(cmd: KeyboardEvent) {
-        if (!cmd.shiftKey) {
-            if (this.gcode !== '') {
-                this.$store.dispatch('printer/sendGcode', this.gcode)
-                this.$store.dispatch('gui/gcodehistory/addToHistory', this.gcode)
-                this.gcode = ''
-                this.lastCommandNumber = null
-                setTimeout(() => {
-                    this.$refs.console.$el.scroll({
-                        top: 0,
-                        left: 0,
-                        behavior: 'smooth',
-                    })
-                }, 20)
-            }
-        } else {
-            this.gcode += '\n'
-        }
-    }
-
-    onKeyUp(): void {
-        if (this.lastCommandNumber === null && this.lastCommands.length) {
-            this.lastCommandNumber = this.lastCommands.length - 1
-            this.gcode = this.lastCommands[this.lastCommandNumber]
-        } else if (this.lastCommandNumber && this.lastCommandNumber > 0) {
-            this.lastCommandNumber--
-            this.gcode = this.lastCommands[this.lastCommandNumber]
-        }
-    }
-
-    onKeyDown(): void {
-        if (this.lastCommandNumber !== null && this.lastCommandNumber < this.lastCommands.length - 1) {
-            this.lastCommandNumber++
-            this.gcode = this.lastCommands[this.lastCommandNumber]
-        } else if (this.lastCommandNumber !== null && this.lastCommandNumber === this.lastCommands.length - 1) {
-            this.lastCommandNumber = null
-            this.gcode = ''
-        }
-    }
-
-    getAutocomplete(e: Event): void {
-        e.preventDefault()
-        if (this.gcode.length) {
-            let check = this.gcode.toLowerCase()
-            const textarea = this.$refs.gcodeCommandField.$refs.input
-            const sentence = textarea.value
-            const len = sentence.length
-            const pos = textarea.selectionStart
-            const currentLinePos = len - reverseString(sentence).indexOf('\n', len - pos)
-            const currentEndPos =
-                sentence.indexOf('\n', currentLinePos) > -1
-                    ? sentence.indexOf('\n', currentLinePos) - 1
-                    : Number.MAX_SAFE_INTEGER
-            if (this.rows > 1) {
-                check = sentence.substr(currentLinePos, currentEndPos - currentLinePos)
-            }
-            let commands = this.helplist.filter((element) => element.commandLow.startsWith(check.toLowerCase()))
-            if (commands?.length === 1) {
-                if (this.rows > 1) {
-                    this.gcode = this.gcode.replace(check, commands[0].command)
-                } else {
-                    this.gcode = commands[0].command
-                }
-            } else if (commands?.length > 1) {
-                let commands = this.helplist.filter((element) => element.commandLow.startsWith(check.toLowerCase()))
-                if (this.rows > 1) {
-                    this.gcode = this.gcode.replace(
-                        check,
-                        commands.reduce((acc, val) => {
-                            return strLongestEqual(acc, val.command)
-                        }, commands[0].command)
-                    )
-                } else {
-                    this.gcode = commands.reduce((acc, val) => {
-                        return strLongestEqual(acc, val.command)
-                    }, commands[0].command)
-                }
-                if (commands && commands.length) {
-                    let output = ''
-                    commands.forEach(
-                        (command) =>
-                            (output +=
-                                '<a class="command font-weight-bold">' +
-                                command.command +
-                                '</a>: ' +
-                                command.description +
-                                '<br />')
-                    )
-
-                    this.$store.dispatch('server/addEvent', { message: output, type: 'autocomplete' })
-                }
-            }
-        }
-        this.$refs.gcodeCommandField.focus()
-    }
-
-    toggleFilter(filter: string) {
-        this.$store.dispatch('gui/updateConsoleFilter', filter)
+        this.gcodeCommandField.setGcode(msg)
     }
 
     mounted() {
@@ -333,20 +142,21 @@ export default class MiniconsolePanel extends Mixins(BaseMixin) {
 
     scrollToBottom() {
         this.$nextTick(() => {
-            if (this.$refs.miniConsoleScroll) {
-                const overlayscroll = this.$refs.miniConsoleScroll.osInstance()
-                overlayscroll?.scroll({ y: '100%' })
-            }
+            this.scrollTo(100)
         })
     }
 
     scrollToTop() {
         this.$nextTick(() => {
-            if (this.$refs.miniConsoleScroll) {
-                const overlayscroll = this.$refs.miniConsoleScroll.osInstance()
-                overlayscroll?.scroll({ y: '0%' })
-            }
+            this.scrollTo(0)
         })
+    }
+
+    scrollTo(position: number) {
+        if (!this.miniConsoleScroll) return
+
+        const instance = this.miniConsoleScroll.osInstance()
+        instance?.scroll({ y: `${position}%` })
     }
 }
 </script>
@@ -358,9 +168,5 @@ export default class MiniconsolePanel extends Mixins(BaseMixin) {
 
 html.theme--light .consoleTable {
     border-top: 1px solid rgba(0, 0, 0, 0.12);
-}
-
-.gcode-command-field {
-    font-family: 'Roboto Mono', monospace;
 }
 </style>
